Use `FileLock` instead. Make test clearer why rebuilding

+++ examples_utils/custom_ops/custom_ops_utils.py
@@ -4,27 +4,15 @@ from time import sleep, time
 
 from cppimport import build_filepath
 import ctypes
-import os
 import shutil
-import sys
-from contextlib import contextmanager
+from contextlib import suppress
+from filelock import FileLock
 
 from examples_utils import sdk_version_hash
+import os
+import logging
 
 __all__ = ['load_custom_ops_lib']
-
-
-@contextmanager
-def open_and_delete(path, mode):
-    """Open file and delete it on exit"""
-    try:
-        with open(path, mode) as f:
-            yield f
-    finally:
-        try:
-            os.remove(path)
-        except IOError:
-            sys.stderr.write('Failed to clean up temp file {}'.format(path))
 
 
 def get_binary_path(path_custom_op: str) -> str:
@@ -51,16 +39,26 @@ def load_custom_ops_lib(path_custom_op: str, timeout: int = 5 * 60):
     t = time()
     while not os.path.exists(binary_path) and time() - t < timeout:
         try:
-            with open_and_delete(lock_path, 'x') as compile_lock:
+            with FileLock(lock_path, timeout=1):
+                if os.path.exists(binary_path):
+                    break
                 cppimport_binary_path = build_filepath(path_custom_op)
                 shutil.copy(cppimport_binary_path, binary_path)
                 os.remove(cppimport_binary_path)
+                logging.debug(f'{os.getpid()}: Built binary')
         except FileExistsError:
+            logging.debug(f'{os.getpid()}: Could not obtain lock')
             sleep(1)
 
     if not os.path.exists(binary_path):
         raise Exception(
-            f'Could not compile binary as lock already taken and timed out. Try deleting the lock: {lock_path}')
+            f'Could not compile binary as lock already taken and timed out. Trying to delete the lock: {lock_path}')
+
+    if os.path.exists(lock_path):
+        with suppress(OSError):
+            os.remove(lock_path)
 
     lib = ctypes.cdll.LoadLibrary(binary_path)
+    logging.debug(f'{os.getpid()}: Loaded binary')
+
     return lib
